pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
name = "birding"
authors = [{ name = "Benned Hedegaard", email = "bhedegaa@cs.brown.edu" }]
description = "Utility scripts and data related to birding."
version = "0.0.1"
classifiers = ["Typing :: Typed"]

readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">= 3.10" # ebird-api requires Python 3.10+

# Core dependencies - Always installed
dependencies = ["ebird-api", "geopy", "jupyter", "python-dotenv", "scipy", "pyinaturalist", "ipyplot", "thefuzz", "pandas"]
# Possibly later: pooch, tqdm, geopandas, pyproj==3.7.0

# Optional dependencies support optional user features
# Reference: https://pydevtools.com/handbook/explanation/what-are-optional-dependencies-and-dependency-groups/
# [project.optional-dependencies]

# Example: Optional dependencies to support computer vision applications
# vision = ["torch", "torchvision", "transformers", "pillow", "opencv-python"]

# all = ["birding[vision]"] # Install everything


# Dependency groups manage development-time dependencies
# Reference: https://pydevtools.com/handbook/explanation/what-are-optional-dependencies-and-dependency-groups/
[dependency-groups]
dev = [
    "ruff",
    "pytest",
    "hypothesis",
    "pytest-cov",
    "mypy",
    "typing-extensions",
]


[tool.hatch.build.targets.wheel]
packages = ["src/birding"] # Include packages at these paths


[tool.ruff]
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = ["ALL"] # Default to all Ruff rules, add ignores as needed
ignore = [
    "D203",   # No blank line between class definition and its docstring
    "D213",   # Place docstring summary immediately after the opening quotes
    "TRY003", # Permit defining exception messages at point of raising
    "EM101",  # Allow using string literals directly within exception constructors
    "EM102",  # Allow using f-strings directly within exception constructors
]
unfixable = ["F401"] # Disable 'fix' for unused imports

# Allow omission of return type on __init__
flake8-annotations = { mypy-init-return = true }

[tool.ruff.lint.per-file-ignores]
"**/tests/*" = ["S101"]   # Allow assertions in test files
"**/scripts/*" = ["T201"] # Allow print statements in scripts
"*.ipynb" = ["T201"] # Allow print statements in Jupyter notebooks


[tool.ruff.format]
quote-style = "double"
indent-style = "space"


[tool.pytest.ini_options]
minversion = "6.0"    # Needed to use pyproject.toml
testpaths = ["tests"] # Directories searched for tests
pythonpath = ["."]    # Added to Python search path


[tool.mypy]
strict = true                   # Enable all optional error checking flags
disallow_incomplete_defs = true # Require all-or-nothing type annotations on functions
disallow_any_unimported = true
implicit_optional = false
check_untyped_defs = true       # Type-check interior of unannotated functions
warn_return_any = true
show_error_code_links = true    # Include documentation links for errors
warn_unused_configs = true      # Warn if per-module mypy settings go unused
